% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anabel-package.R
\docType{package}
\name{anabel-package}
\alias{anabel}
\alias{anabel-package}
\title{anabel: Analysis of Binding Events + l}
\description{
A free software for a fast and easy analysis of 1:1 molecular interaction studies. This package is suitable for a high-throughput data analysis. Both the online app and the package are completely open source. You provide a table of sensogram, tell 'anabel' which method to use, and it takes care of all fitting details. The first two releases of 'anabel' were created and implemented as in (\doi{10.1177/1177932218821383}, \doi{10.1093/database/baz101}).
}
\author{
\strong{Maintainer}: Stefan Kraemer \email{stefan.kraemer.91@gmail.com} (\href{https://orcid.org/0000-0002-0071-9344}{ORCID})

Authors:
\itemize{
  \item Hoor Al-Hasani (\href{https://orcid.org/0000-0002-0431-845X}{ORCID})
  \item Oliver Selinger (\href{https://orcid.org/0000-0001-9723-2809}{ORCID})
}

}
\keyword{internal}
