% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anabel.R
\docType{data}
\name{MCK_dataset_drift}
\alias{MCK_dataset_drift}
\title{Simulated data of binding curve for MCK with linear drift.}
\format{
A data frame with 403 rows and 6 variables:
\describe{
\item{Time}{time points of the binding interaction from start to end}
\item{Conc..50.nM.}{binding curve generated with analyte concentration = 50nM}
\item{Conc..16.7.nM.}{binding curve generated with analyte concentration = 16.7nM}
\item{Conc..5.56.nM.}{binding curve generated with analyte concentration = 5.56nM}
\item{Conc..1.85.nM.}{binding curve generated with analyte concentration = 1.85nM}
\item{Conc..6.17e.1.nM.}{binding curve generated with analyte concentration = 0.617nM}
}
}
\source{
\url{https://apps.cytivalifesciences.com/spr/}
}
\usage{
data(MCK_dataset)
}
\description{
A dataset containing 5 different binding curves of different analyte concentrations with induced baseline drift = -0.01.
Ka = 1e+7nM, Kd = 1e-2
}
\keyword{datasets}
