% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anabel.R
\docType{data}
\name{SCK_dataset_decay}
\alias{SCK_dataset_decay}
\title{Simulated data of different binding curves for SCK method with exponential decay.}
\format{
A data frame with 1091 rows and 6 variables:
\describe{
\item{Time}{time points of the binding interaction from start to end}
\item{Sample.A}{sample containing 5 titerations with Ka = 1e+6nM, Kd = 1e-2}
}
}
\source{
\url{https://apps.cytivalifesciences.com/spr/}
}
\usage{
data(SCK_dataset)
}
\description{
A dataset contains one binding curve with 5 titrations-series (5 injection-series), as follows:
tass: 50, 220, 390, 560, 730;
tdiss: 150, 320, 490, 660, 830;
conc: 6.17e-10 1.85e-09 5.56e-09 1.67e-08 5.00e-08 M
}
\keyword{datasets}
